[Fix][Mamba] Fix prev_states dtype contract in SSDChunkScanFwd
prev_states is produced by SSDChunkStateFwdKernel and
SSDStatePassingFwdKernel as float32 (accum_dtype), but the chunk-scan
kernel declared it as input dtype, making the pipeline incompatible.

- Declare prev_states as accum_dtype (float32) in the TIR kernel
- Cast prev_states -> dtype when loading into state_tile for GEMM
- Generate float32 prev_states in the workload fixture
- Update all docstrings to reflect float32 contract

+++ workloads/mamba.py
@@ -85,7 +85,7 @@ class SSDChunkScanFwdTest(WorkloadBase):
         cb          = torch.randn(b, c, g, L, L, dtype=self.dtype,    device="cuda") * 0.1
         dA_cumsum   = -torch.rand(b, h, c, L,    dtype=torch.float32, device="cuda").cumsum(-1)
         C           = torch.randn(b, S, g, n,    dtype=self.dtype,    device="cuda") * 0.1
-        prev_states = torch.randn(b, c, h, p, n, dtype=self.dtype,    device="cuda") * 0.1
+        prev_states = torch.randn(b, c, h, p, n, dtype=torch.float32,  device="cuda") * 0.1
         dt          = torch.rand( b, h, c, L,    dtype=self.dtype,    device="cuda") * 0.1 + 0.01
         return x, cb, dA_cumsum, C, prev_states, dt
